A new data class that would return the exogenous input series as "x", self history as "y" and learning value as "target". The new data class is to be used with "python_stock_data.pickle"
A new data class that would return the exogenous input series as "x", self history as "y", and target learning value as "target". The new data class is to be used with "python_stock_data.pickle".

Change the tag RET, DATE, PERMNO, VOL as the same of CRSP data base.

Delete the Normalize data file. As the data is directly read from CRSP database.