Another try to reduce memory
The previous push tries to reduce memory usage by replacing the "perceptron" with a linear n-to-1 layer. It also reduces the input size by selecting the stock with the largest trading capital (prc*vol) average per day. These helped a little bit, but not enough.

In this push, the attention weight was assigned as a time independent parameter that does not change at all. Hopefully, it will be optimized between  different epoch. The attention weight is put outside of the time loop. Hopefully, the loop can be rewritten using the lstm instead of the current lstm_cell.